Fix mess corepops report generation made
The corepop report was being injected into the top level readme, not the
one in the corepops directory. I've rectified this and added some checks
to ensure the start/end tags that are used to determine where to inject
the table are present.
Without these the script just repeatedly concatenates the whole readme
to itself on each run... obviously not what we want!