Remove corepops proxy file + dont use supplied corepop
There's no need to use the corepop provided in base, why bother? We
already have a repository of corepops. If we want to include it, then we
should add it to that repository.
We now have a good corepop thanks to Waldek anyway that seems to work
almost everywhere (at least its worked everywhere we've tried).